+++ examples/advanced/src/gen/zod/petController/uploadFileSchema.ts
@@ -9,4 +9,4 @@ export const uploadFileQueryParamsSchema = z.object({ additionalMetadata: z.stri
 /**
  * @description successful operation
  */
-export const uploadFileMutationResponseSchema = z.lazy(() => apiResponseSchema)
+export const uploadFileMutationResponseSchema = z.lazy(() => apiResponseSchema).schema