+++ packages/swagger-zod/src/generators/ZodGenerator.ts
@@ -30,7 +30,17 @@ export class ZodGenerator extends SchemaGenerator<Options, OpenAPIV3.SchemaObjec
     return this
   }
 
-  build({ schema, baseName, description }: { schema: OpenAPIV3.SchemaObject; baseName: string; description?: string }): string[] {
+  build({
+    schema,
+    baseName,
+    description,
+    keysToOmit,
+  }: {
+    schema: OpenAPIV3.SchemaObject
+    baseName: string
+    description?: string
+    keysToOmit?: string[]
+  }): string[] {
     const texts: string[] = []
     const zodInput = this.getTypeFromSchema(schema, baseName)
     if (description) {
@@ -40,7 +50,7 @@ export class ZodGenerator extends SchemaGenerator<Options, OpenAPIV3.SchemaObjec
        */`)
     }
 
-    const zodOutput = zodParser(zodInput, { name: this.options.resolveName({ name: baseName, pluginName }) || baseName })
+    const zodOutput = zodParser(zodInput, { keysToOmit, name: this.options.resolveName({ name: baseName, pluginName }) || baseName })
 
     texts.push(zodOutput)
 
@@ -117,6 +127,10 @@ export class ZodGenerator extends SchemaGenerator<Options, OpenAPIV3.SchemaObjec
         }
         if (schema.format === 'uuid') {
           validationFunctions.push({ keyword: zodKeywords.uuid })
+        }
+
+        if (schema.readOnly) {
+          validationFunctions.push({ keyword: zodKeywords.readOnly })
         }
 
         if (schema.default !== undefined && !Array.isArray(schema.default)) {