+++ packages/react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.11.5](https://github.com/kubb-project/kubb/compare/@kubb/react-v1.11.4...@kubb/react-v1.11.5) (2023-10-16)
+
+
+### 📦 Miscellaneous Chores
+
+* **@kubb/react:** Synchronize undefined versions
+
+## [1.11.4](https://github.com/kubb-project/kubb/compare/@kubb/react-v1.11.3...@kubb/react-v1.11.4) (2023-10-16)
+
+
+### 📦 Miscellaneous Chores
+
+* **@kubb/react:** Synchronize undefined versions
+
 ## [1.11.3](https://github.com/kubb-project/kubb/compare/@kubb/react-v1.11.2...@kubb/react-v1.11.3) (2023-10-14)